Organize routes and folders
