+++ frontend/src/App.js
@@ -1,6 +1,6 @@
 import "./App.scss";
 import { useEffect, useState } from "react";
-import Wishlist from "./routes/WishlistRoute";
+import Wishlist from "./components/wishlists/WishlistRoute";
 
 function App() {
   const [userId, setUserId] = useState(process.env.REACT_APP_USER_ID);
