+++ frontend/src/styles/wishlists/Wishlist.scss
