+++ frontend/src/styles/wishlists/WishlistItem.scss
